refactor: move converters to separate module
