+++ src/sparqlx/utils/utils.py
@@ -1,73 +1,9 @@
 from collections import UserDict
-from collections.abc import Iterable, Iterator
-import json
+from collections.abc import Iterable
 
-import httpx
-from rdflib import BNode, Graph, Literal, URIRef, XSD
 from rdflib.plugins.sparql import prepareQuery
-from sparqlx.utils.types import _TResponseFormat, _TSPARQLBinding, _TSPARQLBindingValue
-
-
-def _convert_bindings(
-    response: httpx.Response,
-) -> list[_TSPARQLBinding]:
-    """Get flat dicts from a SPARQL SELECT JSON response."""
-
-    try:
-        json_response = response.json()
-    except (
-        json.JSONDecodeError
-    ) as error:  # pragma: no cover ; this should be unreachable
-        error.add_note("Note that convert=True requires JSON as response format.")
-        raise error
-
-    variables = json_response["head"]["vars"]
-    response_bindings = json_response["results"]["bindings"]
-
-    def _get_binding_pairs(binding) -> Iterator[tuple[str, _TSPARQLBindingValue]]:
-        """Generate key value pairs from response_bindings.
-
-        The 'type' and 'datatype' fields of the JSON response
-        are examined to cast values to Python types according to RDFLib.
-        """
-        for var in variables:
-            if (binding_data := binding.get(var, None)) is None:
-                yield (var, None)
-                continue
-
-            match binding_data["type"]:
-                case "uri":
-                    yield (var, URIRef(binding_data["value"]))
-                case "literal":
-                    literal = Literal(
-                        binding_data["value"],
-                        datatype=binding_data.get("datatype", None),
-                    )
-
-                    # call toPython in any case for validation
-                    literal_to_python = literal.toPython()
-
-                    if literal.datatype in (XSD.gYear, XSD.gYearMonth):
-                        yield (var, literal)
-                    else:
-                        yield (var, literal_to_python)
-
-                case "bnode":
-                    yield (var, BNode(binding_data["value"]))
-                case _:  # pragma: no cover
-                    assert False, "This should never happen."
-
-    return [dict(_get_binding_pairs(binding)) for binding in response_bindings]
-
-
-def _convert_graph(response: httpx.Response) -> Graph:
-    _format, *_ = response.headers["content-type"].split(";")
-    graph = Graph().parse(response.content, format=_format)
-    return graph
-
-
-def _convert_ask(response: httpx.Response) -> bool:
-    return response.json()["boolean"]
+from sparqlx.utils.converters import _convert_ask, _convert_bindings, _convert_graph
+from sparqlx.utils.types import _TResponseFormat
 
 
 class MimeTypeMap(UserDict):
